apps/configs/views.py
from django_filters.rest_framework import DjangoFilterBackend
from rest_framework import generics, permissions, status
from rest_framework.decorators import action
from rest_framework.filters import OrderingFilter, SearchFilter
from rest_framework.response import Response

from apps.base.mixins import CustomModelViewSet, MultipleSerializerMixin
from apps.base.response import Ok
