requirements/requirements.txt
django>=3.2,<4.0
psycopg2-binary==2.8.5
djangorestframework==3.14.0
django-environ==0.4.5
python-dotenv==0.13.0
drf-yasg==1.21.5
djangorestframework-simplejwt==5.2.2
django-cors-headers==3.11.0
drf-extra-fields==3.4.0
django-filter==21.1
